Diagnostics for derivative sensor
